// ethos-types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
// ...
      tickets: {
        Row: {
          acciones_correctivas: Json | null
          asignado_a: string | null
          created_at: string | null
          departamento:
            | Database["public"]["Enums"]["ticket_departamento"]
            | null
          descripcion: string | null
          estado: Database["public"]["Enums"]["ticket_estado"] | null
          id: number
          media_links: Json | null
          numero_contacto_ticket: string | null
          perfil_cliente_id: number | null
          prioridad: Database["public"]["Enums"]["ticket_prioridad"] | null
          propiedad_id: number | null
          titulo: string | null
          updated_at: string | null
        }
        Insert: {
          acciones_correctivas?: Json | null
          asignado_a?: string | null
          created_at?: string | null
          departamento?:
            | Database["public"]["Enums"]["ticket_departamento"]
            | null
          descripcion?: string | null
          estado?: Database["public"]["Enums"]["ticket_estado"] | null
          id?: number
          media_links?: Json | null
          numero_contacto_ticket?: string | null
          perfil_cliente_id?: number | null
          prioridad?: Database["public"]["Enums"]["ticket_prioridad"] | null
          propiedad_id?: number | null
          titulo?: string | null
          updated_at?: string | null
        }
        Update: {
          acciones_correctivas?: Json | null
          asignado_a?: string | null
          created_at?: string | null
          departamento?:
            | Database["public"]["Enums"]["ticket_departamento"]
            | null
          descripcion?: string | null
          estado?: Database["public"]["Enums"]["ticket_estado"] | null
          id?: number
          media_links?: Json | null
          numero_contacto_ticket?: string | null
          perfil_cliente_id?: number | null
          prioridad?: Database["public"]["Enums"]["ticket_prioridad"] | null
          propiedad_id?: number | null
          titulo?: string | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "tickets_perfil_cliente_id_fkey"
            columns: ["perfil_cliente_id"]
            isOneToOne: false
            referencedRelation: "perfiles_cliente"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "tickets_propiedad_id_fkey"
            columns: ["propiedad_id"]
            isOneToOne: false
            referencedRelation: "propiedades"
            referencedColumns: ["id"]
          },
        ]
      }
    }
// ...
    Enums: {
      configuracion_frecuencia:
        | "Mensual"
        | "Bimestral"
        | "Trimestral"
        | "Semestral"
        | "Anual"
      factura_estado:
        | "Borrador"
        | "PendienteValidacion"
        | "Aprobada"
        | "Enviada"
        | "Pagada"
        | "Vencida"
        | "Anulada"
        | "PagadaConComprobante"
        | "Parcial"
      pago_estado:
        | "PendienteVerificacion"
        | "Verificado"
        | "Rechazado"
        | "Conciliado"
      pago_metodo:
        | "Transferencia"
        | "Cheque"
        | "Efectivo"
        | "TarjetaCredito"
        | "TarjetaDebito"
        | "Otro"
      perfil_cliente_rol: "Propietario" | "Arrendatario" | "Externo"
      perfil_cliente_tipo_persona: "Natural" | "Juridica"
      perfil_operacional_rol: "Jefe Operativo" | "Administrador" | "Directorio"
      propiedad_actividad:
        | "No definida"
        | "Comercio y distribucion"
        | "Venta al por mayor y menor"
        | "Distribucion equipos electronicos"
        | "Almacenaje y dist. alimentos"
        | "Distribucion ropa o textiles"
        | "Servicios de logistica"
        | "Almacenaje y gestion inventarios"
        | "Servicios transporte y distribucion"
        | "E-commerce y envio productos"
        | "Manufactura y ensamblaje"
        | "Ensamblaje productos electronicos"
        | "Fabricacion productos pequenos"
        | "Imprentas y serigrafia"
        | "Carpinteria o fabricacion muebles"
        | "Servicios de tecnologia"
        | "Reparacion equipos electronicos"
        | "Desarrollo software o aplicaciones"
        | "Soporte tecnico e informatica"
        | "Diseno grafico y multimedia"
        | "Oficina administrativa"
        | "Consultoria (financiera, legal, RRHH)"
        | "Agencias de marketing digital"
        | "Gestion de proyectos o eventos"
        | "Servicios contables y auditoria"
        | "Alquiler de espacios"
        | "Alquiler bodegas almacenamiento"
        | "Alquiler oficinas compartidas"
        | "Servicios de impresion"
        | "Impresion gran formato"
        | "Servicios fotocopiado y escaneo"
        | "Impresion material publicitario"
        | "Comercio repuestos o autopartes"
        | "Venta piezas y repuestos vehiculos"
        | "Venta equipos y herramientas esp."
        | "Agencias de seguridad"
        | "Venta y dist. sistemas seguridad"
        | "Instalacion equipos seguridad"
        | "Artes y entretenimiento"
        | "Estudio foto o video"
        | "Taller pintura o escultura"
        | "Produccion eventos o espectaculos"
        | "Servicios reparacion y mantenimiento"
        | "Reparacion electrodomesticos"
        | "Reparacion computadoras"
        | "Mantenimiento maquinaria vehiculos"
        | "Servicios educativos"
        | "Centro formacion o capacitacion"
        | "Clases computacion o diseno"
        | "Talleres y cursos especializados"
        | "Cuidado personal"
        | "Centro estetica o peluqueria"
        | "Gimnasio o centro entrenamiento"
        | "Restauracion y alimentos"
        | "Produccion alimentos empaquetados"
        | "Fabricacion panaderia reposteria"
      propiedad_estado_construccion:
        | "enPlanos"
        | "terreno"
        | "enConstruccion"
        | "obraGris"
        | "acabados"
        | "finalizada"
        | "remodelacion"
        | "demolicion"
        | "abandonada"
        | "paralizada"
      propiedad_estado_entrega: "entregado" | "noEntregado"
      propiedad_estado_uso: "enUso" | "disponible"
      servicio_tipo: "CuotaRecurrente" | "ServicioAdicional" | "Ajuste" | "Otro"
      ticket_departamento:
        | "Consulta"
        | "Administrativo"
        | "Filtraciones"
        | "Fugas de Agua"
        | "Problemas Eléctricos"
        | "Cobranza"
        | "Constructivo"
        | "Limpieza y Mantenimiento"
      ticket_estado: "abierto" | "en_progreso" | "resuelto" | "cerrado"
      ticket_prioridad: "baja" | "media" | "alta"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never